Fix room tracking requests

+++ application/controllers/pos_log.php
@@ -28,31 +28,36 @@ class Pos_log extends User_Controller
 
     public function create()
     {
-
         $data = json_decode(file_get_contents("php://input"));
-        //var_dump($data);
         $user = $this->user_m->get_by(array("UUID" => $data->uuid), true);
         $room = $this->room_m->get_by(array("UUID" => $data->placeId), true);
         $id_action = $this->actual_in_room_m->get_by(array("id_user" => $user->id), true);
-        if ((((count($id_action) != 0 and $data->action == "ENTER")) or (count($id_action) == 0 and $data->action == "LEAVE")) ) {
-            return;
-        }
+       
+        // Dodajemy usera jeśli trzeba
         if (count($user) == 0) {
             $this->user_m->save(array("UUID" => $data->uuid, "nickname" => "Test", "description" => "test"));
+            $user = $this->user_m->get_by(array("UUID" => $data->uuid), true); // TODO: ...
         }
 
-        //var_dump($room);
+        // Dodajemy pokój jeśli trzeba
         if (count($room) == 0) {
             $this->room_m->save(array("UUID" => $data->placeId, "id_category" => 1));
+            $room = $this->room_m->get_by(array("UUID" => $data->placeId), true); // TODO: ...
         }
-        $this->pos_log_m->save(array("id_user" => $user->id, "time" => $data->time, "description" => $data->action, "room" => $room->id_room));
-        if ($data->action == "ENTER") {
-            $this->actual_in_room_m->save(array("id_user" => $user->id, "id_room" => $room->id_room));
-        } else if ($data->action == "LEAVE") {
+        
+        if ($data->action == "LEAVE" && count($id_action) != 0 && $id_action->id_room == $room->id_room) { // Wychodzimy z pokoju w ktorym jesteśmy
             $this->actual_in_room_m->delete($id_action->id_actual);
+            $this->pos_log_m->save(array("id_user" => $user->id, "time" => $data->time, "description" => $data->action, "room" => $room->id_room)); // Zapisujemy wyjście do loga
         }
-        //$this->data['data'] = $data;
-        //$this->load->view('api', $this->data);
+        if ($data->action == "ENTER" && (count($id_action) == 0 || $id_action->id_room != $room->id_room)) { // Wchodzimy do pokoju innego niz obecny (lub nie jesteśmy w pokoju)
+            if(count($id_action->id_actual) != 0) { // Ale już gdzieś jesteśmy i nie dostaliśmy odpowiedniego wyjścia, więc dodamy je sobie sami (może aplikacja padła albo coś)
+                $this->pos_log_m->save(array("id_user" => $user->id, "time" => $data->time, "description" => "LEAVE", "room" => $id_action->id_room)); // Dopisujemy 'sztucznie'
+                $this->actual_in_room_m->delete($id_action->id_actual); // I kasujemy obecność
+            }
+            
+            $this->actual_in_room_m->save(array("id_user" => $user->id, "id_room" => $room->id_room)); // Zapiszmy że jesteśmy w pokoju // TODO: Dodać tu timestampy trzeba!!!
+            $this->pos_log_m->save(array("id_user" => $user->id, "time" => $data->time, "description" => $data->action, "room" => $room->id_room)); // Zapisujemy wejście do loga
+        }
     }
 
     public function get()
